+++ src/components/navbar/index.tsx
@@ -1,8 +1,6 @@
 import React from "react";
-import Link from "next/link";
 import DesktopNav from "./components/desktop";
 import MobileNav from "./components/mobile";
-import { NavBarCSS } from "./styles";
 
 const NavBar = () => {
   return (
